requirements.txt
matplotlib
numpy
opencv-python
Pillow
